Refactor the macros to simplify the crate's API
The `app!` macro together with the `commands` module always felt like an
unnecessarily complicated part of this crate's public API. They both
have been removed in favor of a simpler `main!` macro that sets up the
top-level command for the CLI and runs the application in an async tokio
runtime.
