--- crates/clawless-cli/src/commands.rs
@@ -1,3 +1,0 @@
-pub mod new;
-
-clawless::app!();
