--- examples/hello-world/src/commands.rs
@@ -1,3 +1,0 @@
-mod greet;
-
-clawless::app!();
