+++ examples/hello-world/src/main.rs
@@ -1,7 +1,3 @@
-use clawless::clawless;
+mod greet;
 
-mod commands;
-
-fn main() {
-    clawless!()
-}
+clawless::main!();
